Module attestation::mock

Helpers functions for tests.

Functions are split into internal functions and public functions. Public functions are generic for any runtime, so that they can be reused in other tests. Internal functions/structs can only be used in attestation tests.

Structs

AttestationCreationDetails

MockAccessControl

Authorize iff the subject of the origin and the provided attester id match.

Functions

claim_hash_from_seed

generate_base_attestation

generate_base_attestation_creation_details

generate_base_attestation_with_deposit

insert_attestation

sr25519_did_from_public_key

Type Aliases

Hash