Crate attestation

Attestation Pallet

Provides means of adding KILT attestations on chain and revoking them.

Config

Call

Pallet

Terminology

Claimer:: A user which claims properties about themselves in the format of a CType. This could be a person which claims to have a valid driver’s license.

Attester:: An entity which checks a user’s claim and approves its validity. This could be a Citizens Registration Office which issues drivers licenses.

Verifier:: An entity which wants to check a user’s claim by checking the provided attestation.

CType:: CTypes are claim types. In everyday language, they are standardised structures for credentials. For example, a company may need a standard identification credential to identify workers that includes their full name, date of birth, access level and id number. Each of these are referred to as an attribute of a credential.

Attestation:: An approved or revoked user’s claim in the format of a CType.

Delegation:: An attestation which is not issued by the attester directly but via a (chain of) delegations which entitle the delegated attester. This could be an employe of a company which is authorized to sign documents for their superiors.

Assumptions

The claim which shall be attested is based on a CType and signed by the claimer.

The Verifier trusts the Attester. Otherwise, the attestation is worthless for the Verifier

Re-exports

pub use crate::attestations::AttestationDetails;

pub use crate::default_weights::WeightInfo;

pub use crate::pallet::*;

Modules

attestations

authorized_by

benchmarking

default_weights

Autogenerated weights for attestation

migrations

mock

Helpers functions for tests.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Traits

AttestationAccessControl

Allow for more complex schemes on who can attest, revoke and remove.